Code review feedback: split migrations.
